+++ spec/views/generic_file/new.html.erb_spec.rb
@@ -2,9 +2,9 @@ require 'spec_helper'
 
 describe 'generic_files/new.html.erb', type: :view do
   let(:user_collections) do
-    default_option = SolrDocument.new(id: -1, title_tesim: "Select collection...")
-    col1 = SolrDocument.new(id: "c1", title_tesim: "col1")
-    col2 = SolrDocument.new(id: "c2", title_tesim: "col2")
+    default_option = SolrDocument.new(id: -1, title_tesim: 'Select collection...')
+    col1 = SolrDocument.new(id: 'c1', title_tesim: 'col1')
+    col2 = SolrDocument.new(id: 'c2', title_tesim: 'col2')
     [default_option, col1, col2]
   end
   let(:generic_file) { stub_model(GenericFile, id: '123') }
@@ -16,7 +16,6 @@ describe 'generic_files/new.html.erb', type: :view do
       assign(:batch_id, batch_id)
       assign(:user_collections, user_collections)
       allow(controller).to receive(:current_user).and_return(stub_model(User))
-      allow(GenericFilesController).to receive(:upload_complete_path).with(batch_id).and_return("foo")
       Sufia.config.upload_to_collection = upload_to_collection
     end
 
@@ -27,6 +26,34 @@ describe 'generic_files/new.html.erb', type: :view do
         render
         page = Capybara::Node::Simple.new(rendered)
         expect(page).to have_selector('select#collection', count: 2) # one per tab
+        expect(page).to have_select('collection', options: ['Select collection...', 'col1', 'col2'])
+      end
+
+      context 'and passed a default collection' do
+        context "and the collection is on the user's list of editable collections" do
+          before do
+            assign(:collection_id, 'c2')
+          end
+
+          it 'has default collection selected' do
+            render
+            page = Capybara::Node::Simple.new(rendered)
+            expect(page).to have_select('collection', options: ['Select collection...', 'col1', 'col2'], selected: 'col2')
+          end
+        end
+
+        context "and the collection is NOT on the user's list of editable collections" do
+          before do
+            assign(:collection_id, 'c3_noedit')
+          end
+
+          it 'has instructions selected' do
+            render
+            page = Capybara::Node::Simple.new(rendered)
+            expect(page).to have_select('collection', options: ['Select collection...', 'col1', 'col2'])
+            expect(page).not_to have_select('collection', selected: 'col3_noedit')
+          end
+        end
       end
     end