+++ History.md
@@ -5,6 +5,7 @@
 * Add configuration allowing select menu on batch upload to upload files to a collection. [E. Lynette Rayle]
 * Bump active-fedora version to 9.4 [E. Lynette Rayle]
 * Bump hydra-collections to 5.0.3 [E. Lynette Rayle]
+* Add button 'Upload files' on collection show page which goes to batch upload with collection select menu set to the calling collection [E. Lynette Rayle]
 
 ## 6.3.0